requirements.txt
psycopg==3.1.18
Django==5.0.4
djangorestframework==3.15.1